requirements.txt
aniso8601==9.0.1
blinker==1.8.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Mail==0.10.0
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.10
flask-swagger-ui==4.11.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pymongo==4.8.0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
Werkzeug==3.0.4
